Visualizing Exclusive Lasso Regularization Paths

Visualizing Exclusive Lasso Regularization Paths

# S3 method for ExclusiveLassoFit
plot(
  x,
  xvar = c("norm", "lambda", "l1"),
  label = FALSE,
  legend = "topright",
  ...
)

Arguments

x
An ExclusiveLassoFit object produced by exclusive_lasso

xvar
Value to use on the x-axis (ordinate).

norm: The value of the composite \(\ell_2/\ell_1\) norm "exclusive lasso" norm

lambda: The log of the regularization parameter \(\lambda\).

l1: The \(\ell_1\)-norm of the solution

label
Should variables be labeled? Set label=NULL to label only those variables which are non-zero at the sparsest end of the regularization path.

legend
Should a legend of top variables (heuristically selcted) be displayed? Set legend=NULL to disable, else give the location of the legend.

...
Additional arguments passed to plotting functions n <- 200 p <- 500 groups <- rep(1:10, times=50) beta <- numeric(p); beta[1:10] <- 3

X <- matrix(rnorm(n * p), ncol=p) y <- X

exfit <- exclusive_lasso(X, y, groups) plot(exfit) plot(exfit, legend=NULL, xvar="lambda")